+++ lib/presentation/agent/add-edit_listings/pages/edit_listing.dart
@@ -259,7 +259,7 @@ class EditListing extends GetView<AddListingsController> {
           BaseController().showLoading();
           await Listing(
                   name: controller.propertyNameController.text,
-                  price: controller.propertyCostController.text.toInt(),
+                  price: controller.propertyCostController.text.toDouble(),
                   photos: controller.images,
                   ppsqm: controller.pricePerSqmController.text.toDouble(),
                   // floorArea: controller.floorAreaController.text.toDouble(),
